change sign-up structure
