+++ src/routes/sign-up/components/SignUpProvider.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { GoogleReCaptchaProvider } from 'react-google-recaptcha-v3';
+import SignUpLayout from './SignUpLayout';
+
+const SignUpProvider = () => {
+  return (
+    <GoogleReCaptchaProvider reCaptchaKey={process.env.REACT_APP_CAPTCHA}>
+      <SignUpLayout />
+    </GoogleReCaptchaProvider>
+  )
+};
+
+export default SignUpProvider;
