+++ src/routes/sign-up/index.ts
@@ -1,0 +1,3 @@
+import SignUpProvider from './components/SignUpProvider';
+
+export default SignUpProvider;
